// src/app/modules/asegurado/models/trabajor-existencia.ts
export interface TrabajorExistencia {
  personaId: number
  aseguradoId: number
  estadoAsegurado: number
  estadoGrupoFamiliar: number
  paterno: string
  materno: string
  nombres: string
  matricula: string
  documentoIdentidad: string
  fechaRegistro: Date
  fechaRegistroaseg: Date
  extension: string
  fechaNacimiento: Date
}